+++ packtools/checks.py
@@ -2,13 +2,19 @@
 from __future__ import unicode_literals
 import logging
 import itertools
+import json
 
 import plumber
 
 from .style_errors import StyleError
+from . import catalogs
 
 
 LOGGER = logging.getLogger(__name__)
+
+
+with open(catalogs.ISO3166_CODES) as f:
+    ISO3166_CODES_SET = set(json.load(f))
 
 
 # --------------------------------
@@ -36,7 +42,7 @@ def teardown(message):
 def StyleCheckingPipeline():
     """Factory for style checking pipelines.
     """
-    return plumber.Pipeline(setup, funding_group, doctype, teardown)
+    return plumber.Pipeline(setup, funding_group, doctype, country_code, teardown)
 
 
 # --------------------------------
@@ -132,6 +138,24 @@ def doctype(message):
         err = StyleError()
         err.message = "Missing DOCTYPE declaration."
         err_list.append(err)
+
+    return message
+
+
+@plumber.filter
+def country_code(message):
+    """Check country codes against iso3166 alpha-2 list.
+    """
+    et, err_list = message
+
+    elements = et.findall('//*[@country]')
+    for elem in elements:
+        value = elem.attrib['country']
+        if value not in ISO3166_CODES_SET:
+            err = StyleError()
+            err.line = elem.sourceline
+            err.message = "Element '%s', attribute country: Invalid country code \"%s\"." % (elem.tag, value)
+            err_list.append(err)
 
     return message